refactor(client): unify subnet/IP parsing behind categorizeIpEntry
Extract the token-classification ladder that discovery and the fleet/rack
subnet filters had each reimplemented into one shared core,
`categorizeIpEntry`, and rename networkDiscovery.ts -> ipParsing.ts to match
its now cross-cutting role (it is imported by filters, not just discovery).

Each surface projects the shared result onto its own shape:
- discovery's `parseManualTargets` buckets into scan targets (and still
  rejects IPv6 CIDRs, which it cannot scan);
- the filter helpers (`validateSubnetLine` / `normalizeSubnetLine` /
  `classifySubnetLine`) map onto MinerListFilter's ip_cidrs / ip_ranges and
  reject hostnames.

Behavior is preserved (discovery test suite unchanged); this only removes the
duplicated dispatch. `isValidIpv6Cidr` moves into ipParsing so the core can
use it.
